Began creation loop
